TL2/layout-diff/requirements.txt
python3
python-lxml
python-argh
python-argcomplete
